src/main/java/javaguru/sunday/student_glebs_surtajevs/lesson_5/level_5/ArrayUtil.java
package main.java.javaguru.sunday.student_glebs_surtajevs.lesson_5.level_5;

//Task_31
//Task_32
//Task_33
//Task_34


import main.java.javaguru.sunday.teacher.annotations.CodeReview;

import java.util.Random;

@CodeReview(approved = true)
class ArrayUtil {

    public int[] createArray(int arrayLength) {
        return new int[arrayLength];
    }

    public int[] fillArrayWithRandomNumbers(int[] array) {
        Random random = new Random();
        for (int i = 0; i < array.length; i++) {
            array[i] = random.nextInt(100);
        }
        return array;
    }
}
